+++ examples/contracts/examples/events_with_meta.rs
@@ -17,9 +17,7 @@ abigen!(
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    let client =
-        Provider::<Ws>::connect("wss://mainnet.infura.io/ws/v3/c60b0bb42f8a4c6481ecd229eddaca27")
-            .await?;
+    let client = Provider::<Ws>::connect("wss://eth.llamarpc.com").await?;
 
     let client = Arc::new(client);