+++ examples/events/examples/filtering.rs
@@ -4,9 +4,8 @@ use ethers::{
 };
 use eyre::Result;
 use std::sync::Arc;
-use tokio;
 
-const HTTP_URL: &str = "https://mainnet.infura.io/v3/c60b0bb42f8a4c6481ecd229eddaca27";
+const HTTP_URL: &str = "https://eth.llamarpc.com";
 const V3FACTORY_ADDRESS: &str = "0x1F98431c8aD98523631AE4a59f267346ea31F984";
 const DAI_ADDRESS: &str = "0x6B175474E89094C44Da98b954EedeAC495271d0F";
 const USDC_ADDRESS: &str = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48";
@@ -40,8 +39,7 @@ async fn main() -> Result<()> {
         let tick_spacing = U256::from_big_endian(&log.data[29..32]);
         let pool = Address::from(&log.data[44..64].try_into()?);
         println!(
-            "pool = {}, token0 = {}, token1 = {}, fee = {}, spacing = {}",
-            pool, token0, token1, fee_tier, tick_spacing,
+            "pool = {pool}, token0 = {token0}, token1 = {token1}, fee = {fee_tier}, spacing = {tick_spacing}"
         );
     }
     Ok(())